% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGAnd-deprecated}
\alias{SGAnd-deprecated}
\alias{SGAnd}
\title{Return the AND predicate for use in query construction}
\usage{
SGAnd()
}
\description{
THIS FUNCTION IS DEPRECATED. PLEASE USE THE SGSelectMetaModels FUNCTION
INSTEAD.
}
