% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGConstructOrQuery-deprecated}
\alias{SGConstructOrQuery-deprecated}
\alias{SGConstructOrQuery}
\title{Produce a Query with all the values separated by OR operators.}
\usage{
SGConstructOrQuery(vValues, fieldName)
}
\arguments{
\item{vValues}{a vector of values}

\item{fieldName}{the name of the field}
}
\value{
a character string that can be used with the SGQuery function
}
\description{
THIS FUNCTION IS DEPRECATED. PLEASE USE THE SGSelectMetaModels FUNCTION
INSTEAD.
}
