% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGContains-deprecated}
\alias{SGContains-deprecated}
\alias{SGContains}
\title{Returns the Contains term for use in query construction}
\usage{
SGContains(field, value)
}
\arguments{
\item{field}{one of the fields listed by the SGGetMetaModelQueryFields function}

\item{value}{the value for this field

THIS FUNCTION IS DEPRECATED. PLEASE USE THE SGSelectMetaModels FUNCTION
INSTEAD.}
}
\description{
This corresponds to the LIKE SQL operator which is case sensitive.
#'
}
