% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGContains}
\alias{SGContains}
\title{Returns the Contains term for use in query construction
Note : This corresponds to the LIKE SQL operator which is case sensitive}
\usage{
SGContains(field, value)
}
\description{
Returns the Contains term for use in query construction
Note : This corresponds to the LIKE SQL operator which is case sensitive
}
