% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGFilterMetaModels}
\alias{SGFilterMetaModels}
\title{Searches the metamodel database for matches to the specified query}
\usage{
SGFilterMetaModels(query = NULL)
}
\arguments{
\item{query}{a string containing the query.  Null will return all metamodels in the database.
Queries should be created using the
query utility methods.

example : SGQuery(SGContains("mmid", "FMU02664"), SGAnd(), SGContains("outputType", "Coniferous"))

Note : Groups / associativity is not supported by queries.  To perform a query of type :
(TermA AND TermB) OR TermC you will have to create a query for (TermA AND TermB),
then a second query for TermC and then combine the results together.}
}
\value{
a dataframe containing all metamodels matching the query
}
\description{
Searches the metamodel database for matches to the specified query
}
\seealso{
SGQuery, SGContains, SGGreaterThan, SGLowerThan, SGAnd, SGOr
}
