% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGGetFinalSample}
\alias{SGGetFinalSample}
\title{Provide the Final Sample of Parameter Estimates.}
\usage{
SGGetFinalSample(mmid)
}
\arguments{
\item{mmid}{A string containing the mmid of the MetaModel}
}
\value{
a data.frame object
}
\description{
The final sample is the one provided by a selection
rate of the Markov chain, This function does not rely
on a cache on the server end. The full meta-model must
be deserialized and this may take a few seconds.
}
