% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGGetMetaData}
\alias{SGGetMetaData}
\title{Provide the Meta-Data of a particular Meta-Model.}
\usage{
SGGetMetaData(mmid)
}
\arguments{
\item{mmid}{A string containing the mmid of the meta-model}
}
\value{
a list object
}
\description{
Provide the Meta-Data of a particular Meta-Model.
}
