% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGGetMetaModelFieldCombinations}
\alias{SGGetMetaModelFieldCombinations}
\title{Return the field combinations and the count of each combination.}
\usage{
SGGetMetaModelFieldCombinations(fields)
}
\arguments{
\item{fields}{a vector of field names}
}
\value{
a data.frame object
}
\description{
Return the field combinations and the count of each combination.
}
