% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGGetMetaModelQueryFields}
\alias{SGGetMetaModelQueryFields}
\title{Return the available fields of the metamodels in the database that can be
used in queries.}
\usage{
SGGetMetaModelQueryFields(nchar = 100)
}
\arguments{
\item{nchar}{the number of characters to be kept in the 'values' field (100 by default)}
}
\value{
a data.frame object
}
\description{
The possible values are also returned.
}
