% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGGetRegenerationLagIfAny}
\alias{SGGetRegenerationLagIfAny}
\title{Provide the regeneration lag if any.}
\usage{
SGGetRegenerationLagIfAny(mmid)
}
\arguments{
\item{mmid}{A string containing the mmid of the MetaModel#'}
}
\value{
a numeric
}
\description{
Provide the regeneration lag if any.
}
