% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGGetSummary}
\alias{SGGetSummary}
\title{Provide the summary of a particular Meta-Model.}
\usage{
SGGetSummary(mmid)
}
\arguments{
\item{mmid}{A string containing the parameter estimates and
other information.}
}
\value{
a list object
}
\description{
Provide the summary of a particular Meta-Model.
}
