% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGGreaterThan}
\alias{SGGreaterThan}
\title{Returns the GreaterThan term for use in query construction}
\usage{
SGGreaterThan(field, value)
}
\arguments{
\item{field}{one of the fields listed by the SGGetMetaModelQueryFields function}

\item{value}{the value for this field}
}
\description{
Returns the GreaterThan term for use in query construction
}
