% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGLowerThan}
\alias{SGLowerThan}
\title{Returns the LowerThan term for use in query construction}
\usage{
SGLowerThan(field, value)
}
\description{
Returns the LowerThan term for use in query construction
}
