% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGOr-deprecated}
\alias{SGOr-deprecated}
\alias{SGOr}
\title{Returns the OR predicate for use in query construction}
\usage{
SGOr()
}
\description{
THIS FUNCTION IS DEPRECATED. PLEASE USE THE SGSelectMetaModels FUNCTION
INSTEAD.
}
