% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGPredict}
\alias{SGPredict}
\title{Gets a set of predictions from the specified metamodel}
\usage{
SGPredict(mmid, ageyrmin, ageyrmax, step = NULL, varout = NULL)
}
\arguments{
\item{mmid}{A string containing the mmid of the MetaModel#'}

\item{ageyrmin}{The minimum age year to get the prediction set for}

\item{ageyrmax}{The maximum age year to get the prediction set for (ageyrmax will be included in output values)}

\item{step}{(optional int) The number of years to use from ageyrmin to ageyrmax for predictions (default is 1 if not specified)}

\item{varout}{(optional) A string specifying the variance output mode from the following :
NONE (default) : no variance output
PARAMEST : variance output for parameter estimation
PARAMESTRE : variance output for parameter estimation including random effect}
}
\value{
a data.frame object with two columns. "age" is the age of the stand (yr) and "v" is the prediction (e.g. m3/ha)
}
\description{
Gets a set of predictions from the specified metamodel
}
\seealso{
SGPredictMC
}
