% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGPredictMC}
\alias{SGPredictMC}
\title{Gets Predictions using Monte-Carlo Simulated Parameters.}
\usage{
SGPredictMC(mmid, ageyrmin, ageyrmax, step = NULL, nbsub = 1, nbreal = 1)
}
\arguments{
\item{mmid}{A string containing the mmid of the MetaModel}

\item{ageyrmin}{The minimum age year to get the prediction set for}

\item{ageyrmax}{The maximum age year to get the prediction set for (ageyrmax will be included in output values)}

\item{step}{(optional int) The number of years to use from ageyrmin to ageyrmax for predictions (default is 1 if not specified)}

\item{nbsub}{The number of subjects to simulate parameters for in range (0,1000). Zero disables the
variability due to the random effect.}

\item{nbreal}{The number of realizations to simulate parameters for in range (0,1000). Zero disables
the variability due to the parameter estimates.}
}
\value{
a data.frame object with four columns (RealizationID, SubjectID, AgeYr, Pred).
}
\description{
The function simulates deviates for the parameter estimates. The variability of the
random effects and the parameter estimates can be disabled by setting the nbsub and
nbreal arguments to zero, respectively.
}
\seealso{
SGPredict
}
