% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGQuery-deprecated}
\alias{SGQuery-deprecated}
\alias{SGQuery}
\title{Constructs a query using the provided terms and predicates
A query is a series of one or more terms separated by predicates.
Example : mmid contains "foo" AND outputType contains "All" OR nbRealizations > 1000}
\usage{
SGQuery(...)
}
\arguments{
\item{...}{the multiple operators of the query

THIS FUNCTION IS DEPRECATED. PLEASE USE THE SGSelectMetaModels FUNCTION
INSTEAD.}
}
\description{
Constructs a query using the provided terms and predicates
A query is a series of one or more terms separated by predicates.
Example : mmid contains "foo" AND outputType contains "All" OR nbRealizations > 1000
}
