% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGSelectMetaModels}
\alias{SGSelectMetaModels}
\title{Provide a Selection of Meta-Models}
\usage{
SGSelectMetaModels(
  geoDomain = NULL,
  dataSource = NULL,
  climateChangeOption = NULL,
  growthModel = NULL,
  outputType = NULL,
  fitModel = NULL,
  stratumGroup = NULL
)
}
\arguments{
\item{geoDomain}{the geographic domain of the meta-model (e.g. the
bioclimatic subdomain in Quebec)}

\item{dataSource}{the source of data used to fit the meta-model (e.g. PET4
which is the fourth campaign of the provincial inventory in Quebec)}

\item{climateChangeOption}{the climate change scenario}

\item{growthModel}{the complex growth model used to run the simulation
(e.g. artemis2009)}

\item{outputType}{the response variable (e.g. AliveVolume_AllSpecies)}

\item{fitModel}{the formulation of the meta-model (e.g. ChapmanRichards)}

\item{stratumGroup}{the name of the stratum group that is represented
by the meta-model}
}
\value{
a data.frame object
}
\description{
Each field is queried using OR as logical operator between the
different values. The queries are then linked using an AND operator.
}
\details{
For instance, if the geoDomain argument is the vector c("6OUEST", "6EST")
and the dataSource argument is the vector c("PET3", "PET4"), then all the
meta models fitted in either geoDomain 6OUEST or 6EST and fitted to either
dataSource PET3 or PET4 are selected.
}
\examples{
myMetaModels <- SGSelectMetaModels(geoDomain = c("6OUEST", "6EST"), dataSource = c("PET3", "PET4"))

}
\seealso{
the SGGetMetaModelQueryFields function for the values of the different
fields
}
