% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGStatus}
\alias{SGStatus}
\title{Provide the Status of the Web API.}
\usage{
SGStatus()
}
\value{
a data.frame object with four columns (RealizationID, SubjectID, AgeYr, Pred).
}
\description{
Provide the Status of the Web API.
}
