% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGContains}
\alias{SGContains}
\title{Returns the Contains term for use in query construction
This corresponds to the LIKE SQL operator which is case sensitive.}
\usage{
SGContains(field, value)
}
\arguments{
\item{field}{one of the fields listed by the SGGetMetaModelQueryFields function}

\item{value}{the value for this field}
}
\description{
Returns the Contains term for use in query construction
This corresponds to the LIKE SQL operator which is case sensitive.
}
