% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGFindBest}
\alias{SGFindBest}
\title{Find the Metamodel(s) that Best Fit a Particular Context.}
\usage{
SGFindBest(geoRegion, geoDomain, outputType, ecoType, leadingSpecies = NULL)
}
\arguments{
\item{geoRegion}{the geoRegion (e.g., QC)}

\item{geoDomain}{the geoDomain (e.g., 2OUEST)}

\item{outputType}{the outputType (e.g., Volume)}

\item{ecoType}{the ecotype (e.g., MS22)}

\item{leadingSpecies}{an option character string that stands
for the leading species (e.g., EN)}
}
\value{
a data.frame instance
}
\description{
The function can process multiple queries if all the arguments
are of the same length.
}
