% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGGetMetaModelQueryFields}
\alias{SGGetMetaModelQueryFields}
\title{Returns the available fields of the metamodels in the database that can be
used in queries}
\usage{
SGGetMetaModelQueryFields()
}
\value{
a list of field names
}
\description{
Returns the available fields of the metamodels in the database that can be
used in queries
}
