% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGGOFGraph}
\alias{SGGOFGraph}
\title{Provide a Goodness-of-Fit Graph for a Particular MetaModel.}
\usage{
SGGOFGraph(mmid, textsize = 20, plotPred = T, title = mmid, ymax = 250)
}
\arguments{
\item{mmid}{A string containing the mmid of the MetaModel}

\item{textsize}{the font size (20 by default)}

\item{plotPred}{a boolean true to enable the plot of predicted values}

\item{title}{an optional title for the graph (the mmid by default)}

\item{ymax}{the upper bound of the y axis (250 by default)}
}
\value{
a ggplot2 graph
}
\description{
The graph displays the simulations and the predictions of
the metamodel instance.
}
