% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGLowerThan}
\alias{SGLowerThan}
\title{Returns the LowerThan term for use in query construction}
\usage{
SGLowerThan(field, value)
}
\arguments{
\item{field}{one of the fields listed by the SGGetMetaModelQueryFields function}

\item{value}{the value for this field}
}
\description{
Returns the LowerThan term for use in query construction
}
