% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGPredictMC}
\alias{SGPredictMC}
\title{Gets predictions using Monte-Carlo simulated parameters}
\usage{
SGPredictMC(mmid, ageyrmin, ageyrmax, step = NULL, nbsub = 1, nbreal = 1)
}
\arguments{
\item{mmid}{A string containing the mmid of the MetaModel}

\item{ageyrmin}{The minimum age year to get the prediction set for}

\item{ageyrmax}{The maximum age year to get the prediction set for (ageyrmax will be included in output values)}

\item{step}{(optional int) The number of years to use from ageyrmin to ageyrmax for predictions (default is 1 if not specified)}

\item{nbsub}{The number of subjects to simulate parameters for in range (1,1000)}

\item{nbreal}{The number of realizations to simulate parameters for in range (1,1000)}
}
\value{
a data.frame object with four columns. "real" is the realization id
"sub" is the subject id,  "age" is the age of the stand (yr) and "v" is the prediction (e.g. m3/ha)
}
\description{
Gets predictions using Monte-Carlo simulated parameters
}
\seealso{
SGPredict
}
