% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSStandGrowth4R.R
\name{SGStatus}
\alias{SGStatus}
\title{Provide the Status of the Web API.}
\usage{
SGStatus()
}
\value{
a list containing information on the Web API (e.g. version).
}
\description{
Provide the Status of the Web API.
}
